Challenge/lv3 14

+++ src/main/java/org/example/expert/domain/todo/dto/response/TodoSearchResponse.java
@@ -10,10 +10,10 @@ public class TodoSearchResponse {
   private final int managerCount;
   private final int commentCount;
 
-  public TodoSearchResponse(Todo todo, long managerCount, long commentCount) {
+  public TodoSearchResponse(Todo todo, int managerCount, int commentCount) {
     this.todoId = todo.getId();
     this.title = todo.getTitle();
-    this.managerCount = (int) managerCount;
-    this.commentCount = (int) commentCount;
+    this.managerCount = managerCount;
+    this.commentCount =  commentCount;
   }
 }
